social_media_api/pyproject.toml
[project]
name = "social-media-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "decouple>=0.0.7",
    "django>=5.2.9",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
]
